pyproject.toml
[project]
name = "weather-web-app"
version = "0.1.0"
description = "A simple weather app using Streamlit"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit",
    "requests",
    "python-dotenv",
]
